# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MLOVI - Multidisciplinar: BACKEND API V1",
        "description": "Description of your API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Main (production) server"
        }
    ],
    "paths": {
...
        "/especialidades": {
            "get": {
                "tags": [
                    "Especialidades"
                ],
                "summary": "Listar todos os especialidades",
                "operationId": "getEspecialidadeList",
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida"
                    },
                    "401": {
                        "description": "Não Autenticado"
                    },
                    "403": {
                        "description": "Proibido"
                    }
                }
            },
            "post": {
                "tags": [
                    "Especialidades"
                ],
                "summary": "Cadastrar um novo especialidade",
                "description": "Cria um novo registro de especialidade com os dados fornecidos.",
                "operationId": "storeEspecialidade",
                "requestBody": {
                    "description": "Dados do especialidade",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "cbos"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Psicologia"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Descrição da especialidade"
                                    },
                                    "cbos": {
                                        "type": "string",
                                        "example": "2515-10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Especialidade criado com sucesso"
                    },
                    "400": {
                        "description": "Erro na requisição"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Proibido"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/especialidades/{id}": {
            "get": {
                "tags": [
                    "Especialidades"
                ],
                "summary": "Buscar especialidade por ID",
                "operationId": "getEspecialidadeById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do especialidade",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida"
                    },
                    "401": {
                        "description": "Não Autenticado"
                    },
                    "403": {
                        "description": "Proibido"
                    },
                    "404": {
                        "description": "Especialidade não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Especialidades"
                ],
                "summary": "Atualizar especialidade",
                "operationId": "updateEspecialidade",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do especialidade",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados do especialidade a serem alterados. Verifique os campos no Schema",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Psicologia"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida"
                    },
                    "401": {
                        "description": "Não Autenticado"
                    },
                    "403": {
                        "description": "Proibido"
                    },
                    "404": {
                        "description": "Especialidade não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Especialidades"
                ],
                "summary": "Excluir especialidade",
                "operationId": "deleteEspecialidade",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do especialidade",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Especialidade excluído com sucesso"
                    },
                    "401": {
                        "description": "Não Autenticado"
                    },
                    "403": {
                        "description": "Proibido"
                    },
                    "404": {
                        "description": "Especialidade não encontrado"
                    }
                }
            }
        },
...
    },
...
}